+++ atmosphere/cAtmosphereModel.h
@@ -157,7 +157,7 @@ private:
     }
 
     std::vector<std::vector<int> > i_topography;
-    std::vector<int> tropopause_layers; // keep the tropopause layer index
+    std::vector<double> tropopause_layers; // keep the tropopause layer index
     std::vector<double> temp_tropopause; // lateral temperature distribution along the tropopause
     void init_tropopause_layers();
